Generalized linear model (abstract class)

This class specializes ParamRegrModel. It is the abstract base class for parametric generalized linear model objects with specific distribution family such as NormalGLM and handles the (inverse) link function.

Super class

gofreg::ParamRegrModel -> GLM

Methods

Public methods

GLM$new()

GLM$mean_yx()

GLM$clone()

Inherited methods

gofreg::ParamRegrModel$F1_yx()

gofreg::ParamRegrModel$F_yx()

gofreg::ParamRegrModel$f_yx()

gofreg::ParamRegrModel$fit()

gofreg::ParamRegrModel$get_params()

gofreg::ParamRegrModel$sample_yx()

gofreg::ParamRegrModel$set_params()

Method new()

Initialize an object of class GLM.

Usage

GLM$new(linkinv = identity, params = NA)

Arguments

linkinv

inverse link function, defaults to identity function

params

model parameters to use as default (optional)

Method mean_yx()

Evaluates the regression function or in other terms the expected value of Y given X=x.

Usage

GLM$mean_yx(x, params = private$params)

Arguments

x

vector of covariates

params

model parameters to use, defaults to the fitted parameter values

Returns

value of the regression function

Method clone()

The objects of this class are cloneable with this method.

Usage

GLM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.