Negative log-likelihood function for a parametric regression model

The log-likelihood function for a parametric regression model with data (x,y) is given by the sum of the logarithm of the conditional density of Y given X=x evaluated at y.

This function is one option that can be used to fit a ParamRegrModel. It returns the negative log-likelihood value in order for optim() to maximize (instead of minimize).

Usage

loglik_xy(data, model, params)

Arguments

data

list() with tags x and y containing the data

model

ParamRegrModel to use for the likelihood function

params

vector with model parameters to compute likelihood function for

Value

Value of the negative log-likelihood function