Parametric resampling scheme for a paramatric regression model under random censorship

Generate a new, resampled dataset of the same shape as data following the given model. The covariates X are kept the same. Survival times Y are drawn according to model$sample_yx() and censoring times C according to the KM estimator.

Usage

resample_param_cens(data, model)

Arguments

data

data.frame() with columns x, z and delta containing the original data

model

ParamRegrModel to use for the resampling

Value

data.frame() with columns x, z and delta containing the resampled data

Examples

# Create an example dataset
n <- 10
x <- cbind(runif(n), rbinom(n, 1, 0.5))
model <- NormalGLM$new()
params <- list(beta = c(2, 3), sd = 1)
y <- model$sample_yx(x, params = params)
c <- rnorm(n, mean(y) * 1.2, sd(y) * 0.5)
z <- pmin(y, c)
delta <- as.numeric(y <= c)
data <- dplyr::tibble(x = x, z = z, delta = delta)

# Fit the model to the data
model$fit(data, params_init = params, inplace = TRUE, loglik = loglik_xzd)

# Resample from the model given data
resample_param_cens(data, model)
#> # A tibble: 10 × 3
#>     x[,1]  [,2]       z delta
#>     <dbl> <dbl>   <dbl> <dbl>
#>  1 0.918      0  1.29       1
#>  2 0.0509     1 -0.0582     1
#>  3 0.513      1  3.64       1
#>  4 0.757      1  4.09       0
#>  5 0.580      0  1.63       1
#>  6 0.0846     1  2.23       0
#>  7 0.735      1  3.26       0
#>  8 0.131      0  1.27       1
#>  9 0.864      0  0.958      1
#> 10 0.677      0  2.69       1