Parametric resampling scheme for a paramatric regression model with resampling of covariates

Generate a new, resampled dataset of the same shape as data following the given model. The covariates are resampled from data$x and the response variables are drawn according to model$sample_yx().

Usage

resample_param_rsmplx(data, model)

Arguments

data

data.frame() with columns x and y containing the original data

model

ParamRegrModel to use for the resampling

Value

data.frame() with columns x and y containing the resampled data

Examples

# Create an example dataset
n <- 10
x <- cbind(runif(n), rbinom(n, 1, 0.5))
model <- NormalGLM$new()
params <- list(beta = c(2, 3), sd = 1)
y <- model$sample_yx(x, params = params)
data <- dplyr::tibble(x = x, y = y)

# Fit the model to the data
model$fit(data, params_init = params, inplace = TRUE)

# Resample from the model given data
resample_param(data, model)
#> # A tibble: 10 × 2
#>     x[,1]  [,2]       y
#>     <dbl> <dbl>   <dbl>
#>  1 0.720      0  1.74  
#>  2 0.208      1  4.40  
#>  3 0.758      1  6.66  
#>  4 0.705      1  6.13  
#>  5 0.645      1  5.38  
#>  6 0.335      1  2.75  
#>  7 0.989      1  5.06  
#>  8 0.0286     1  0.764 
#>  9 0.475      0  2.31  
#> 10 0.225      0 -0.0101